examples/keystonemiddleware/requirements.txt
python-glanceclient==2.6.0
keystonemiddleware==4.15.0
oslo.config==4.1.1
## The following requirements were added by pip --freeze:
Babel==2.4.0
PyYAML==3.12
appdirs==1.4.3
argparse==1.2.1
debtcollector==1.13.0
funcsigs==1.0.2
functools32==3.2.3-2
iso8601==0.1.11
jsonpatch==1.15
jsonpointer==1.10
jsonschema==2.6.0
keystoneauth1==2.20.0
monotonic==1.3
msgpack-python==0.4.8
netaddr==0.7.19
netifaces==0.10.5
openstacksdk==0.9.16
os-client-config==1.27.0
oslo.i18n==3.15.3
oslo.serialization==2.18.0
oslo.utils==3.25.1
pbr==3.0.1
positional==1.1.1
prettytable==0.7.2
pyparsing==2.2.0
python-keystoneclient==3.10.0
pytz==2017.2
requests==2.14.2
requestsexceptions==1.2.0
rfc3986==1.0.0
six==1.10.0
stevedore==1.21.0
urllib3==1.21.1
warlock==1.3.0
wrapt==1.10.10
wsgiref==0.1.2
## The following requirements were added by pip --freeze:
WebOb==1.7.2
oslo.context==2.14.0
oslo.log==3.26.1
pycadf==2.5.0
pyinotify==0.9.6
python-dateutil==2.6.0
## The following requirements were added by pip --freeze:
deprecation==1.0.1
